fzf(man_widget): execute man/help directly
when the fzf-man-widget widget is triggered (pressing alt-h),
now it would behave similary like zsh run-help builtin.

though, wherein If the buffer is empty, it executes <cmd> man pages.
In case there is no manual entry, it fallbacks and display --help on a bat pager.
Otherwise, it opens the fzf widget.

Andalso, some minor improvements, on the widget query.
Instead of obtaining the entire value of the buffer,
it now specifically extracts only the last word from it.